rings: use depth buffer
So that we don't render the rings on top of the planets.

We have to be a bit careful with the way it works: we compute the depth
to be around the position of the planet we render, so it is not a
global range for the whole scene like in normal OpenGL rendering.  This
mean the depth range cannot be computed in the screen projection
function, so we have to do it 'manually' in the renderer.

All of this is still a bit experimental.