Add first draft of kepler orbit for planets
It's almost totally untested.  I just added Himalia (Jupiter moon) in
the ini file with an orbit taken from HORIZONS, just to check that the
code doesn't crash.

The code should be pretty easy to understand: if a planet defines an
orbit as a kepler element list in the ini file we use that to compute
the position.

I added a script that will try to update the orbit elements in the ini
file with up to date values from HORIZONS.  Unfortunately only using a
single element value is not precise as soon as we move a few weeks after
the element date.  Eventually I would like to have some way to bundle
more elements in the data so that we don't have to update every few
days.

The planet update function is turning into a mess, it need a good
cleanup, but for the moment I mostly want to test the accuracy of the
kepler orbit with real data and put all the solar system objects in the
ini file.