Generate the stars data from python
Instead of doing it in the core.

This allow to clean up some of the C code, since we don't have to
support generating eph files.

It is easier to use python for all the data generation.

For some reason the actual data is different, I guess this is due to the
compression, or to floating point number precision.  I hope this doesn't
change anything.