Hips relax a bit the render order
After we compute the theoretical float render order, we truncate to the
closest rounded value, instead of the next integer.  That way the
landscape isn't rendered at level 1 when we don't zoom.

Shouldn't make much difference for the other surveys, maybe the image
won't be pixel perfect sometimes.  If that becomes a problem we could
add a parameter so that we can specify the rendering quality we want.

For the moment this is needed to prevent our cloudfront server cost to
get too high!