Hips: low res rendering only use allsky texture
If we try to render a survey at very low resolution and we have an
allsky image, we don't even try to load the order zero textures.

I changed the API of hips a little bit to use flags so that we can pass
the HIPS_FORCE_ALLSKY flag along for this.

This is done so that we don't hit the data server too much.