Remove progressbars after a while
I changed the API of progressbar util so that the bars can automatically
be removed from the list if the report function is not called for a while.

I also stop printing out the progress bars in the C version, instead I
added the list in the gui debug tab.