Fix moon contrast
The current survey used has a too high contrast.

I guess it would be better to fix the survey directly, but I think it
could be useful to be able to adjust the contrast of any of the textures
so I added a uniform for that in the shader.

For the moment we hard code the contrast adjustment for the moon.