Support light speed correction
It works by recomputing the position twice: once at the current time to
get the physical position, and once at the time corrected to take into
account the light travel.

I use it by default on all the planets.  It might affect the
performances thought, so maybe we should only do it for certain planets.

Also, I am not sure if this should be done to the Sun as well!