Planet: split the update function in two
We have to be a bit careful, because obj_update will compute the
position taking into account light speed, but this is not always what we
want (for parent offset we need to physical position).

In planets.c, planet_update_ will update the internal pvg values,
without touching the obj.pos.pvg, and so we can use it when we need the
physical position.  Not sure this is the best way to do it.  Maybe the
observer should have two TT time values: one for the actual time, and
one for the light speed corrected time?