Correct for light speed in planets.c
I was hoping I could do it in obj.c, but this is a bit complicated
because sometimes we need the non corrected position, and sometimes the
corrected on.

For the moment I reverted the previous algo, and instead I use the
object speed to adjust for the light speed directly in planets.c.  We
store both positions in the planet object as pvg and pvgc, so that we
can decide which one to use.