Can delete tile with non running loaders
This fix a bug when some tiles started a loading worker but then were
not rendered anymore.  Since the loader didn't get destroyed the cache
wouldn't evict the tile, and after a while the cache would be filled
with those tiles.

We could maybe add a callback function in the main loop to do the
cleanup here?