Batch render the lines
The idea is that we maintain a list of deferred rendering items.  If we
can we try to reuse the last item in the list when we add lines.

The goal would be to use rend item for all the rendering commands, so
that we can do the rendering only after all the commands have been
issued.