Don't add gaia stars that we already have
Some stars are both in the original bundled catalog (from hipparcos) and
the gaia online sources.

Eventually we should probably use the exact same survey and just
pre-bundle some of the tiles, but for the moment we just discard gaia
stars if the vmag is lower than the highest bundled star vmag.