Add some table support in eph files
I want to change the format so that we adding new columns into the data
doesn't require to change the code.

This means that we cannot use simple binupack to load the data.  Instead
we tell the file loader what columns we care about and the loader
automatically parse them, ignoring the rest.

For the moment it is not working totally yet.