eph-file: let the client do the chunk parsing
That is the eph format is just a container of chunks.  But we provide
helper functions to parse common data structures within the chunks.

This makes the code a bit easier I think, and it allows some
optimizations: for example we don't have to copy the compressed data
twice in stars.c.