Use a link preload tag to load the WASM early
Changed the way the WASM is loaded so that the emscripten module is passed
directly a wasm binary files rather than letting it load the remote file.
This allows to initiate download right from the start and speeds up first
sky display by a few seconds.

It was necessary modify the preload-webpack-plugin to force crossorigin
option to the link tag.