assets: add 'asset_release' function
This is important not to keep asset data that we don't use anymore around.

The original plan was to allow the assets manager to decide when to
delete the data, but there are some cases when we know for sure that we
won't access the same asset anymore and then it makes sense to destroy it
as soon as possible for memory limited environments.