Stars: don't show names for gaia stars
The problem is that some stars are still in both the HIP catalog and the
Gaia one.  For example Polaris is HIP 11767 and Gaia DR2
576402619921505664.

The trick we use to skip gaia stars with vmag lower than the min vmag of
the hip bundled data doesn't work, because in gaia survey, the bmag is
very high (about 10).  So both entries appears in the sky, and the name
is rendered twice).

The solution could be to add proper vmag in the gaia survey.  But until
then I just prevent rendering names for gaia stars.