core->fov is the min fov in x and y
This is what people would expect when they set the fov.

In order to keep the projection init function clear, they all take
argument in fovx (fov in the x direction).

I added a function 'projection_compute_fovs' that can be used to compute
the fovs in x and y given the min fov.