Add precision argument to orbit algo
The Kepler equation solution used is not good enough for eccentric
orbits of comets.

For the moment I allow to set the precision to zero to use the old style
approximation.  I think we should probably use the proper Kepler loop
all the times, but I need to check if the performances are good enough.