obj_get does not use canonical form
Because I want obj_get to work with comets id, like 'C/1 1995 O1', and
the canonical form would break that as 'C/ 1 1995 O 1'.

This is not backward compatible, but I think should work the same for
all cases that we care about.  The actual search algo is going to be
outside of the core anyway.