Adjuste points size depending on smoothness
The goal is that if we change the smoothness of the stars and planet
point (that is the halo effect around them) they sill seem to be more or
less the same size.

The size we give to a point should correspond to the size it would have
with no smoothness effect at all.  With smoothness the size should be
the size of where we feel the point border is.  I used a factor of
1/(1 - 0.7 * smoothness) that seems to work.

The actual value for the '0.7' could be computed by integrating the
shader function.  If the smoothness effect was linear, we should have
0.5, since the shader use a sqrt, I adjusted this value a bit.