Add obj_add_data_source function
Not used yet.  The idea is that the modules won't automatically load
data, instead at init time we load all the data sources we want and each
module that recognise a source type will be able to use it.

This will give more flexibility to use different sources on different
platforms.