Add sources module
Simple module that handle the case when we want to add a data source
that is a directory containing a list of data source.

This is the common case for things like landscapes or skycultures.  Each
module is able to add a single source directory, but in case of a list
it's easier to have a common code for all sources.

Created as a module so that it doesn't mess with the core too much.