Sources module handle hipslist sources
If we add a source with the type 'hipslist', the sources manager will
first download and parse the list, then the properties files, and then
will query the modules to see wich one can handle this kind of hips
survey.

This was previously done directly in the planets module, but this allow
to move the source data logic outside of the module.