Prevent splitting moon tile at large fov
The problem is that the algorithm that computes the number of split to
use in the tile assume we render at the order that correspond to the
resolution.  But for planets viewed from far away, even order zero is
too large.

We could use some kind of logical orders that go below zero maybe.  For
the moment I hardcode the split to 4 in that case.

This improve the rendering speed.