Add texture->border attribute
This can be used to consider the last pixel of a texture as outside the
UV of the texture, for better interpolation on the sides.

It would be better to be able to set the actual sub texture size or a UV
mapping, but this is a bit more complicated and for the moment we just
need to support a single pixel border.