paint_points take radius in pixel instead of angle
This is a first step to prepare for the new tone mapping code.  The
function to compute point radius will return a radius in pixel directly,
which is more like what an HDR rendering pipeline would do.

For the moment I added a convertion function so that we can still use
the old core_get_point_for_mag code.