Use new tonemapper
Logarithmic algo instead of the previous Tumblin alogo.  This one seems
to work better for the atmosphere.

I also make the core tonemapper a direct structure instead of a pointer,
so that we can directly access its attributes.

I tried to adjust the values to make the rendering look ok, but we
probably need more tweaking.