Add a live profiler with remotery: https://github.com/Celtoys/Remotery
To use it we have to compile with 'make remotery'.  See the doc of
remotery for how to see the profiling.

I added a small wrapper on top of remotery to make the code a bit
cleaner.  We have to add the PROFILE macro on top of all the function we
want to monitor.  I already added a few.