Make healpix clip test stricter
The previous algo would often considers that tiles at low level
behind us are visible, because the stereographic projection doesn't
really work with clipping test in that case.

I used an heuristic that hopefully works: if all the points of a tile
are behind us, and none are visible on the screen, then we assume the
tile is not visible.

Might need to find a better way to do that.