Constellation supports art anchors UV in pixels
So that we can parse the original stellarium format for constellation
images.

I don't think it's a good idea to store the coordinates in pixels,
because then we can't change the images resolution without also having to
change the anchors values.  But it's nice to have direct support for
stellarium format.