Labels module checks oid to differentiate labels
Since we support labels fade-in, we have to verify if a label we render
is already on screen.  Only testing for the label value doesn't work if
we render several labels with the same text at different positions (like
with the grid).

This commit add a test on the oid so that we can render several labels
with same text while still having the fade-in effect.