Fonts have a scale value
Nanovg seems to use the wrong sizes for fonts: for example the stars
label should have a size of 13px, but the rendered font doesn't match
what we get with other program with a 13px DejaVuSans font.

It seems to depend on the font used, so I added a argument to the
sys_list_fonts fonction so that at least we can give a scale associated
with each font.