Try to fix stars get_by_oid method
The previous algo would only search the gaia stars if the search was
done using a gaia id.  But the new online gaia survey also has HIP
values for the stars, and so a non gaia oid doesn't necessary that the
stars is not from the gaia survey!

Probably need more cleanup on this code.