Fix bug with previous stars commit
We have to be careful not to traverse the gaia survey at high fov, since
the min order is 3, so it only make sens to start when we are already
quite zoomed in.

This is achieved here by making sure we check the survey min vmag value
before we call hips_traverse.