Remove constellation set_image attribute function
Instead we do a direct C call across modules.

I want to remove all attributes from sky objects.  This one could have
been replaced by a method on the constellations module, but since for
the moment we only use it from the skycultures module, it's easier to do
it like that.