Hips only support allsky at level zero
This is the only case where they are actually useful (to prevent loading
all the 12 level zero tiles of planets when they are very small), and it
makes the code simpler like that.

In fact I could probably remove allsky support totally, and handle it
manually in the planet module?