Fix lunar eclipse with scaled moon
I make it work by also scaling the shadow spheres.  This seems to work
properly for the moon, but I am not sure if this is totally correct in
general.

I guess an other way would be to pass the scale in the shader and use it
for everything except the shadow computation.