Attempt to fix bug with lines
Need to make sure we split the line enough to have a smooth rendering,
and to avoid clipping error.

A proper fix would be for the clipping test to check for culling, and
also to compute the split of individual lines.  I do it like that for
the moment since this doesn't change the code.