Fix scaled planets shadow casting
The problem is that the previous solution was to scale the shadow casting
with the same factor as the planet.  This doesn't work properly.

The new approach here is to only use the scaling for the projected
position in the shader, not for the model position.

I think this should fix the problem.