Make obj_t stateless (BIG CHANGE!)
This is a big refactorization of the way we access sky objects
attributes, like position and magnitude.  Instead of doing `obj_update`
and attribute access, we now access the infos with `obj_get_info`, that
takes an observer as argument.  If a module wants to cache the computed
values for an observer it has to do it itself.

- Remove the obj_update function.  For modules we add the module_update
  function that doesn't take an observer as argument.

- Add obj_get_info, with an associated obj_t method, that each object
  class needs to implement for all the supported informations (some
  might still be missing at the moment!)

- Remove pvo and vmag attributes from obj_t: each object class is free
  to implement those the way it wants.  We just need to make sure that
  the get_info method returns the correct value.

- Remove the INFO macro and all the info properties in the classes.  We
  still support PROPERTY for the modules.  Sky object infos now should
  only be retrieved with obj_get_info, we cannot set them.

- Update all the modules to use the new mechanism.

This most likely introduce some regression.