Some fix in the debug page
All the layers need to have a proper id.  This is actually a bit
annoying: we need that so that we can render the json tree of the
modules, but I guess having a array would also work.  Probably need to
change the format of the json tree to fix that.