Major atmosphere and tone mapping update
The atmosphere now looks more realistic and its luminance
is visually calibrated against real observations (made on
August 3rd in France).
It required a rewrite of the GLSL shader, and a different
implemenation of the sky luminance model.

The update also allowed to suppress some ad-hoc rendering
settings but more of those remain.