Fix crash when we change skyculture
If a constellation is selected.  The problem is that in that case the
constellation get unparented, but the core still attempt to update it.

This means the update function should not assume the object still has a
parent.

We should probably automatically deselect any object whose only ref is
the core selection & lock though.