Improves layer_get_by_oid
So that if we add an object to a layer that support the get_by_oid
function we call it as well.

This is needed for the geojson support: we could have a layer containing
geojson object, each of them also containing features.  We should be
able to search for a given feature by oid.