Fix planet quad culling test
We can't do the usual culling test with a non linear projection, because
a surfaces behind the observer would have a normal pointing in the
opposite direction as the view, while still being visible.

To fix that I do the test with a dot product using the center of the
planet as reference.  Let see if that is enough...