Add obj_get_pos convenience function
To replace the less generic obj_get_pos_observed, since we will soon
have to use a new frame (FRAME_MOUNT) for some operations that were done
in the observed frame before.

Also the previous implementation of obj_get_pos_observed always
considered the ICRF position to be at infinity, which was not correct!