Simplify projections discontinuity code
We only support a discontinuity at the azimuth = 180° semi circle, and
we don't support splitting projections into two parts anymore.

For the moment we just set a flag to tell if the projection has a
discontinuity and we let the painter compute the intersection itself
instead of using methods of the projections.