Geojson: add function to get clicked features
Not used or even tested yet.  This will only be available in the js
interface for the moment.

Since we are now doing the mesh projection on the GPU, we use the 2d area
buffer like we do for the rest of the objects.  Eventually we should
probably move all the modules to this kind of code.