Extra test for line discontinuity
We check to see if a line is discontinuous directly in the tesselator:
if two points are too far apart from each other we consider the line
discontinuous.

This fix some bugs with discontinuous projections.  In theory we can now
remove the test from the painter.

There might be a better way to do that...