Grid: change the way we render great circles
We render them as a grid with most lines masked.  That we we can use the
same recursion algo.  Previous code was not working well when we start
to zoom in.

It would actually be nicer to fix the paint_line function to support all
sizes, but this will do for the moment.