Add support for new skyculture json format
We want to merge all the sky culture data files into a single json file,
so that we can extend it with extra information, and separate the
description for each constellation/feature.

This is still experimental, and we fallback to the old format if the
index.json file is not present in the sky culture folder.

The format itself will probably change a bit.