Fix bug in planet rendering
Since we use an union in the painter for the planet, atmosphere, and
line options, but the line options are set by default, we need to clear
it to make sure we don't get random values.

We could maybe remove the union?  Or don't set the line attributes by
default?