Revert jcon_parse behavior with undefined values
In a recent commit I changed it so that undefined values are set to zero
automatically, but this poses some problem if the json value is actually
zero, then we have no way to test if it was undefined or not.

I revert to the previous behavior of not changing the variable if they
are undefined.

This should fix a bug in the satellites parsing.

Would be nice to have a way to add a compile time check that all the
variables have been initialized.