Fix core_vmag_from_radius for extrem cases
For example if we zoom on the Moon, the actual vmag that should be
returned should be lower than the initial range used for the dichotomy.

Instead of showing the 'too many iteration' error, we just return the
min value directly.

If we need to get a proper value even for those cases, we could replace
the dichotomy for an algorithm that can extend outside the initial
range.  Newton wouldn't work though, because the mag -> radius function
is clamped at zero.