Add support for tasks in the core
This is a simple mechanism to add and remove functions to be executed
at each frame.

Not used yet.

I add this mostly so that I can remove the 'sources' module and instead
let each module load it's source.  Since that requires some async
loading of data, I will use tasks to make the code simpler to follow.