Stars: some more code cleanup
Still preparing to support any number of stars surveys.  Need to be
careful because the code currently make a lot a assumption about the
fact that we use one offline and one online surveys.

For the moment I just cleanup the code a bit to avoid relying on surveys
index.