Progressbar: better api for 'keepalive' attribute
We need to support progress bars that disappear after a while if there are
no activities, for example for the stars surveys if we move we might
cancel the current progress bars.  The simple way to do that is to set
a timeout after which we remove the bar.

This commit makes it easier to support the other kind of progress bar,
where we don't want a timeout, but instead we remove the progress bar
only after completion.