Allow to use fallback fonts with nanovg
In js, I renamed 'setFont' to 'addFont', we can now call it with
different fonts and the following ones will be used as fallback.  We can
also call this with the bundled font paths
(asset://fonts/NotoSans-Regular.ttf and asset://fonts/NotoSans-Bold.ttf)
so that they can be added as last fallback.