DSO: encode source index in oid
Since otherwise we can't ensure that the oid are actually unique.

This fixes a bug where we would select the Small Magellanic Cloud when
we click on the Large Magellanic Cloud.

At some point I think it would be nice to remove the oid completely and
instead make objects creation so fast that we could just create all the
objects of the visible tiles on the fly.