Cleanup eph file units
Make sure that units that have the same dimensions have the same 16
bits.  This allows to check the dimension of the data.  For example if
the data in the eph file is in EPH_RAD, we can request it in EPH_ARCSEC,
but not int EPH_VMAG.

Also add a check (an assert for the moment), that the units we request match
the units in the data.

Added a fix for the gaia survey that is using legacy arcsec unit.