Compute apparent position when loading star
It includes proper motion and parallax. In the longer run we
may want to recompute the effect of proper motion and parallax
at each frame rather than only once.

Warning: we currently assume a catalog epoch of 1991.25, this
need to be corrected in the code once we have the epoch for each
star.