Fix regression in planetary computation
The previous commit fixed the missing aberration effect,
but introduced a regression by assuming a linear space motion
for the planets when compensating for light speed.

This assumption is good for most planets, but not for distant
satellites such as Jovian satellite which need to use a better
model. I.e. in 10 minutes which is the time that the light of
Io comes to earth, its motion cannot be approximated by a
linear motion withotu introducing errors.