Optimise minor planets memory layout
We split the memory so that we can quickly iterate on the minor planets
and discard those that are not flagged as visible.  For that we need
the base structure to be as small as possible.

An other possible way would have been to use a second list for flagged
planets.