Stars: fix skipped stars when loading survey
This test to skip loading stars was added because the Gaia survey
contains stars that might already be in other added surveys.  To prevent
two surveys from having the same stars I added this test on the min_vmag
value.

The actual problem is that the min_vmag value used in the HiPS property
is not exactly correct.

For the moment I fix the problem by only testing for the Gaia survey.