module_list_objs doesn't take an observer
The observer was used to test the max vmag.  But this is probably better
to just let the max vmag be a hint, and let the caller check the actual
magnitude if needed.

This is mostly so that we can list all the potentially bright satellites
without having to compute their position, since we can estimate the
maximum brightness.