Sats: add a fast function to compute the altitude
We will need to compute the satellites altitude very quickly at
different times for the 'tonight' page.  Doing it the proper way has too
much overhead because observer_update is currently quite slow.

I add this experimental function that doesn't require a fully updated
observer.  Not to be used in normal case.