Geojson: support different surveys with same url
This is not something I was planning to do with HiPS surveys, but we
need that since geojson surveys uses filters, and we might want two
surveys of the same url with different filters.

The proper way would be to separate the HiPS source of data from the
computed style attributes, so that we can properly cache the tiles while
still allowing several views with different filters.