Do not pass dt to core_update
Instead we let the core compute it itself from the system clock.

Mostly doing this so that we can let the core the current real time
clock when we start an animation, so that it works even if the core is
not updated.

An other reason is so that we can let the core enforce the
synchronisation of the sky time with the real time clock.