Asteroid orbits properly use J2000 ecliptic
The orbit are based on the J2000 ecliptic, not the JNow.  I put a
statically computed rotation matrix in the code to replace the use of
observer->re2i.

Note that it would be even better to convert the orbit to CIRF, so that
we don't have to do any rotation.