Compute projection at flush time with proper depth
The issue is that we do not know what depth range we should use until we
have all the rendering items.  So the idea is to never project anything
until flush time.  Here we compute the final projection that should be
used for the item that require depth (like the planets).