Fix planet HiPS order computation (again)
The bug I though I fixed in the last commit was in fact due to a typo in
the code (used vec2_norm instead of vec3_norm).

This revert the previous wrong fix and add some comments to make the API
a bit more clear regarding the angles we use (diameter and not radius).