Change Moon artificial scale computation
Previously we only considered the view fov, but this doesn't make sense
if we can travel outside of Earth, since if we get close the Moon, we
don't want to scale it up, even if we have a large fov.

Also add a property to disable the Moon scaling totally.