Try to fix depth range again
This is tricky because we want to support very large ranges, like being
close to the ISS, with the Sun visible.

The previous idea of limiting the near range to keep a good resolution
doesn't work in that case.

I also changed the default depth function to LEQUAL (instead of LESS),
so that the for very far objects, when the precision is too low and all
the depth values are maxed, at least we show objects by rendering order.