Cache allows to set a grace period for items
This is needed for the HiPS cache, so that we can get over the max
cache capacity in case where we really need to render many tiles on
screen.  This happens when we zoom on the moon and change the time very
quickly.

This probably increase the memory usage the app a little bit.