Attempt to fix minor label flickering
When two labels overlaps, sometime they flicker on screen.  I think this
was due to the sort algorithm trying to compare two labels at infinity
where the distances are more or less the same.

This add a test for this case to make sure the comparison function
returns 0 in that case.