constellations: refactor constellation update code
The star loading procedure is now done only once in a single step.
This avoids intermediary state where e.g. stars linked to the
constellation lines are loaded but not the stars linked to the images
anchors.
The new code also properly handles constellation without lines (just
an image) and also single-stars constellations.